support more keybinds and remove some duplicates
Added 'ctrl-/' and 'ctrl-_' as undo
unbinded 'ctrl-x' which caused problems
Moved normal navigation keys in emacs to 'body' so that they work in the command palette and other pop-ups too.
'ctrl-x k' closes the pane, not all of the panes now.

Changed 'ctrl-x ctrl-f' to use fuzzy finder. This is because the normal open-file dialogue is optimized to use mouse, but fuzzy finder works better with a keyboard.
The best solution is to install the "advanced open file" package which works similarly to "ctrl-x ctrl-f" in emacs.

Added 'ctrl-c ctrl-c' to comment/uncomment code.